fix: move target app logic out of gcir loop (#339)
we are still seeing large latency coming from
get_client_initialize_response. Especially in one case, we are passing
in an sdk key with the css-prod target app that only has 33 flags, so we
didn’t expect this method to take several hundred ms to execute.
we are using the ruby sdk.
i checked the code for get_client_initialize_response , it is looping
through every FF and as expected it is bypassing the flag eval if the
flag doesn’t have the correct target app (based on this
[line](https://github.com/statsig-io/ruby-sdk/blob/baf41b80d22d818cb6be3230a6e79e95d47c6782/lib/client_initialize_helpers.rb#L32))
one improvement that 
[@jlai](https://statsigconnect.slack.com/team/U01T8B1H466)
suggested is to move this
[line](https://github.com/statsig-io/ruby-sdk/blob/baf41b80d22d818cb6be3230a6e79e95d47c6782/lib/client_initialize_helpers.rb#L29)
which is converting the sdk key to target app id outside of the loop, so
that the it is only executed once.