feat: loosen typed numeric check (#349)
The `get_typed` method now converts between Integer and Float when
appropriate

Previously, for example, json would only be able to store "0" for a 0
numeric parameter. If you used the typed getter, it was a bit unexpected


context:
exp_config.get_typed("aaa", 11.1)
=> returns 11.1 

exp_config.get("aaa", 11.1)
=> returns 0

get_typed returns the default value if the type of the default value and
the type of the parameter do not match. Its trying to say - you expected
to get a double from this call, but the value is not a double, so its
better for you to get the matching type.
0 is probably a special case here, where it gets treated as an int, and
doesnt match the float type passed in. I think JSON probably just treats
0 as only an integer