[release] 2.4.1 - bugfix - split event logs into max batch size chunks (#375)
We've observed an issue in production where event batches grow well
beyond the `logging_max_buffer_size` in `StatsigOptions`. In the worst
case, they get so large they are rejected by the statsig backend and not
retried.

The fix in this version is to split the full queue into
`logging_max_buffer_size` chunks that are sent in multiple requests

Its possible that under high qps loads the sdk will still not be able to
keep up





>Included In This Release
>- adbd162320610d08551a2e57b7fac49816e45967 tore-statsig
>   - fix: chunk logevent requests into batch_size (#374)