improve performance and memory usage in HashUtils.djb2
Runtime profiling indicates this method generates several many memory allocations.

Comparing to the JS implementation, we saw the intent of the `hash &= hash` line was to force the JS runtime to keep the number as a 32-bit integer. This is indeed the correct way to do it in JS, but not in Ruby; as a result, the `hash` local will grow ever larger, requiring more and more memory since Ruby supports unbounded integers.

Fix: truncate the hash value on each iteration with the same 32-bit `0xFFFFFFF` constant used at the end instead.